bind patient service to billing service and next add modify patient as kafka producer

+++ billing-service/src/main/java/com/koriebruh/billingservice/dto/BillingRequestDTO.java
@@ -14,7 +14,7 @@ public class BillingRequestDTO {
 
     @NotNull
     @Min(value = 1, message = "Patient ID must be greater than 0")
-    private int patientId;
+    private long patientId;
 
     @NotBlank(message = "cannot be blank")
     @Size(min = 1, message = "minimum 1 character")
